rust/ffi: move AppLayerTxData to ffi
Ticket: 7666
