+++ rust/Cargo.toml.in
@@ -95,3 +95,4 @@ htp = { package = "suricata-htp", path = "./htp", version = "@PACKAGE_VERSION@" 
 
 [dev-dependencies]
 test-case = "~3.3.1"
+suricata-ffi = { path = "./ffi", version = "@PACKAGE_VERSION@", features = ["testing"] }
