+++ rust/ffi/Cargo.toml.in
@@ -11,3 +11,4 @@ suricata-sys = { path = "../sys" }
 [features]
 debug = []
 debug-validate = []
+testing = []
